% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barplot_sscs}
\alias{barplot_sscs}
\title{Create a barplot of either count or count fraction by size class}
\usage{
barplot_sscs(x, what = c("counts", "fraction")[2])
}
\arguments{
\item{x}{x tibble of one sscs record}

\item{what}{either 'count' or 'fraction'}
}
\value{
ggplot2 object
}
\description{
Create a barplot of either count or count fraction by size class
}
