% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{compute_crop_by_class}
\alias{compute_crop_by_class}
\title{Compute crop per size class}
\usage{
compute_crop_by_class(x, vf = volume_factor())
}
\arguments{
\item{x}{SSCS class object}

\item{vf}{table of volume conversion factors}
}
\value{
numeric vector of crop available per bin
}
\description{
Compute crop per size class
}
