% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{counts_to_bins}
\alias{counts_to_bins}
\title{Convert a table of counts (by plot) to size bins}
\usage{
counts_to_bins(x, drop_spat = TRUE)
}
\arguments{
\item{x}{table of counts}

\item{drop_spat}{logical, if TRUE drop the 0-10 spat class}
}
\value{
a table of bin size, counts and fractions
}
\description{
Convert a table of counts (by plot) to size bins
}
