% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{draw_map.default}
\alias{draw_map.default}
\title{Draw map of the provided object}
\usage{
\method{draw_map}{default}(x)
}
\arguments{
\item{x}{an object to map}
}
\description{
Draw map of the provided object
}
