% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{export_example}
\alias{export_example}
\title{Export the example}
\usage{
export_example(filename = "softshell_example.xls")
}
\arguments{
\item{filename}{syring, the filename (including path) to save the example to.}
}
\value{
logical (invisibly) with TRUE meaning success
}
\description{
Export the example
}
