% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{export_template}
\alias{export_template}
\title{Export the template}
\usage{
export_template(filename = "softshell_template.xls")
}
\arguments{
\item{filename}{syring, the filename (including path) to save the template to.}
}
\value{
logical (invisibly) with TRUE meaning success
}
\description{
Export the template
}
