% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{gather_sscs}
\alias{gather_sscs}
\title{Gather the stats for a particular SSCS record(s)}
\usage{
gather_sscs(x, n = 1, fields = sscs_classes(prefix = "f"))
}
\arguments{
\item{x}{tibble of sscs records}

\item{n}{the record to gather}

\item{fields}{the fields to gather}
}
\value{
a tibble of class and value suitable for plotting.
}
\description{
Gather the stats for a particular SSCS record(s)
}
