% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_csv_report}
\alias{import_csv_report}
\title{Import Soft Shell Clam Survey data from a simple compound CSV}
\usage{
import_csv_report(
  filename = system.file("examples/softshell.csv", package = "softshell")
)
}
\arguments{
\item{filename}{the name of the file to import}
}
\value{
SSCS S3 object (a list of sscs data)
Comprised of up to 5 items: 
\itemize{
\item{filename}
\item{survey, a tibble of surveydata - required}
\item{bins, a tibble of counts (and fractions) by size bin - optional}
\item{plots, a tibble of plot locations - optional}
\item{counts, a tibble of plot counts - optional}
}
}
\description{
Import Soft Shell Clam Survey data from a simple compound CSV
}
