% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dmr_report}
\alias{import_dmr_report}
\title{Import Soft Shell Clam Survey data from the "Report" sheet of an excel document}
\usage{
import_dmr_report(
  filename = system.file("Yarmouth-Lanes-Is-2019-09-21.xls", package = "softshell")
)
}
\arguments{
\item{filename}{the name of the file to import}
}
\value{
SSCS S3 object (a list of sscs data)
Comprised of up to 5 items: 
\itemize{
\item{filename}
\item{survey, a tibble of surveydata - required}
\item{bins, a tibble of counts (and fractions) by size bin - optional}
\item{plots, a tibble of plot locations - optional}
\item{counts, a tibble of plot counts - optional}
}
surveydata is comprised of the following 
\itemize{
 \item{date survey date}
 \item{town survey town}
 \item{site survey site name}
 \item{lon  decimal degrees longitude (west is negative)}
 \item{lat  dcimal degrees latitude}
 \item{plot_size sample plot size, typically 1' x 2'}
 \item{plot_interval distance between sample plots}
 \item{plot_count number of sample plots}
 \item{area survey coverage in acres}
 \item{count count of softshell clams collected}
 \item{count_legal count of soft shell clams of legal harvest size}
 \item{countf_legal fraction of soft shell clams of legal harvest size}
 \item{abundance abundance of soft shell clams in sample area}
 \item{abundance_legal abundance of soft shell clams of legal size in sample area}
 \item{crop extrapolated abundance of soft shell clams at site}
 \item{crop_legal extrapolated abundance of soft shell clams of leagl size at site}
 \item{cropf_legal extrapolated fractional abundance of soft shell clams of leagl size at site}
}
}
\description{
Import Soft Shell Clam Survey data from the "Report" sheet of an excel document
}
