% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_report_xls}
\alias{import_report_xls}
\title{Import a user generated single-sheet XLS softshell survey file}
\usage{
import_report_xls(
  filename = system.file("examples/softshell-example.xls", package = "softshell"),
  sheet = 1
)
}
\arguments{
\item{filename}{string, the name of the file to read}

\item{sheet}{character or index, the name of the sheet within the file
By default read the first encountered.}
}
\value{
list of 
\itemize{
\item{meta list of metadata}
  \describe{
    \item{date}{Date, survey date}
    \item{town}{character, survey town}
    \item{site}{character, survey site within town}
    \item{crew}{character, crew names}
    \item{plot_interval}{two element numeric, (shore-normal, along-shore)}
    \item{plot_size}{generally character,  1'x2'}
    \item{plot_count}{numeric, the number of plots sampled}
    \item{area}{numeric, aggregate area surveyed in acres}
    \item{class_abundance}{named numeric vector, of abundance by class size}
    \item{abundance}{numeric, total bushels per acre}
    \item{legal_abundance}{numeric, legally sized bushels per acre}
  }
\item{plots tibble of plot locations}
\item{counts tibble of counts at each plot}
\item{abund tibble of abundance (bushels per acre) by size class at each plot}
}
}
\description{
Import a user generated single-sheet XLS softshell survey file
}
