% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_xls_report}
\alias{import_xls_report}
\title{Import a user generated single-sheet XLS softshell survey file}
\usage{
import_xls_report(
  filename = system.file("examples/softshell-example.xls", package = "softshell")
)
}
\arguments{
\item{filename}{string, the name of the file to read}
}
\value{
list of 
\itemize{
\item{meta list of metadata}
\item{plots tibble of plot locations}
\item{counts tibble of counts at each plot}
\item{abund tibble of abundance (bushels per acre) by size class at each plot}
}
}
\description{
Import a user generated single-sheet XLS softshell survey file
}
