% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SSCS}
\alias{plot.SSCS}
\title{Plot softshell abundances as a histogram}
\usage{
\method{plot}{SSCS}(x = SSCS(), ylim = NULL, title = NULL)
}
\arguments{
\item{x}{a SSCS object}

\item{ylim}{either NULL (autoselect) or a specified range}

\item{title}{either NULL or character.  If NULL the title is generated automatically}
}
\value{
ggplot object
}
\description{
Plot softshell abundances as a histogram
}
