% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bins}
\alias{plot_bins}
\title{Plot softshell bins as a histogram}
\usage{
plot_bins(x = SSCS(), title = NULL)
}
\arguments{
\item{x}{a SSCS object}

\item{title}{either NULL or character.  If NULL the title is generated automatically}
}
\value{
ggplot object
}
\description{
Plot softshell bins as a histogram
}
