% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_volume_factor}
\alias{plot_volume_factor}
\title{Plot the volume factor by size class}
\usage{
plot_volume_factor(x = volume_factor())
}
\arguments{
\item{x}{the volume factor table}
}
\value{
ggplot2 plot object
}
\description{
Plot the volume factor by size class
}
