% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R, R/io.R
\name{read_sscs}
\alias{read_sscs}
\title{Read a soft shell clam survey table stored in CSV format}
\usage{
read_sscs(filename = system.file("softshell.sscs", package = "softshell"))

read_sscs(filename = system.file("softshell.sscs", package = "softshell"))
}
\arguments{
\item{filename}{the name of the file to read}
}
\value{
tibble of sscs data

sscs class object
}
\description{
Read a soft shell clam survey table stored in CSV format

Read a sscs format text file
}
