% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softshell-package.R
\name{softshell}
\alias{softshell}
\alias{softshell-package}
\title{Soft Shell Clam Survey Data from Maine Department of Marine Resources}
\description{
Maine Department of Marine Resources [DMR](https://www.maine.gov/dmr/)
publishes reports on soft shell clam surveys.  Methods provide access to example data
from [Yarmouth, Maine](https://yarmouth.me.us), as well as plotting and mapping
tools.
}
