% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscs.R
\name{sscs_classes}
\alias{sscs_classes}
\title{Generate a list of soft shell clam size classes}
\usage{
sscs_classes(prefix = "", sep = "_")
}
\arguments{
\item{prefix}{the characters with which to optionally prefix the classes}

\item{sep}{separator to between prefi and class name}
}
\value{
character vector of class names
}
\description{
Generate a list of soft shell clam size classes
}
