% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscs.R
\name{sscs_units}
\alias{sscs_units}
\title{Retrieve the units for a given variable (if known)}
\usage{
sscs_units(x = "crop_legal")
}
\arguments{
\item{x}{character, the name of the variable}
}
\value{
the units for the specified variable
}
\description{
Retrieve the units for a given variable (if known)
}
