% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{volume_factor}
\alias{volume_factor}
\title{Retrieve the volume factor by size class}
\usage{
volume_factor()
}
\value{
table (tibble) of size_class and vol_factor
}
\description{
Retrieve the volume factor by size class
}
