% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_sscs}
\alias{write_sscs}
\title{Write a sscs dataset}
\usage{
write_sscs(x, filename = "softshell.sscs")
}
\arguments{
\item{x}{a sscs dataset}

\item{filename}{the path and name of the destination file}
}
\description{
Write a sscs dataset
}
