% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscs.R
\name{SSCS}
\alias{SSCS}
\title{Generates S3 class 'SSCS' object form a list of metadata, plot locations, spat and counts}
\usage{
SSCS(x = import_report_xls())
}
\arguments{
\item{x}{a 4 element list
\itemize{
  \item{meta, a list of meta data}
  \item{plots, a data.frame (hopefully tibble) of plot locations}
  \item{spat, a data.frame (hopefully tibble) of spat counts at each plot}
  \item{counts, a data.frame (hopefully tibble) of counts by size class at each plot}
}}
}
\value{
a sscs class object which is a list of 5 elements.  The three inputs
  (with meta slightly augmented) and an additional abund element which is the 
  counts data frame transformed into abundance (bushels per acre).
}
\description{
Generates S3 class 'SSCS' object form a list of metadata, plot locations, spat and counts
}
